Align vc cli with did cli
